requirements.txt
fastapi==0.111.0
fastapi_sessions==0.3.2
bcrypt
pandas
pyotp==2.9.0
python-dotenv==1.0.1
passlib==1.7.4
pydantic==2.7.1
PyJWT==2.8.0
requests
SQLAlchemy==2.0.30
uvicorn==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1